+++ src/modules/youtube/services/index.ts
@@ -1,0 +1,3 @@
+export * from "./playlistItems.service";
+export * from "./search.service";
+export * from "./videos.service";
